+++ fortnite_wiki/fortnite_wiki/Models/Details/WeaponsDetails.swift
@@ -8,7 +8,7 @@
 
 import Foundation
 
-class WeaponsDetails: NSObject {
+class WeaponsDetails: NSObject, Decodable {
 
     var detailLevel: Int
     var weaponId: Int